chore: move plan-generator glue in a separate package
