+++ router/cmd/plan-generator/main.go
@@ -4,20 +4,20 @@ import (
 	"flag"
 	"log"
 
-	"github.com/wundergraph/cosmo/router/cmd"
+	"github.com/wundergraph/cosmo/router/pkg/plan_generator"
+)
+
+var (
+	executionConfigFilePath    = flag.String("execution-config", "config.json", "execution config file location")
+	sourceOperationFoldersPath = flag.String("operations", "operations", "source operations folder location")
+	plansOutPath               = flag.String("plans", "plans", "output plans folder location")
+	operationFilterFilePath    = flag.String("filter", "", "operation filter file location which should contain file names of operations to include")
+	timeout                    = flag.String("timeout", "30s", "timeout")
+	concurrency                = flag.Int("concurrency", 0, "how many query plan run concurrently")
+	help                       = flag.Bool("help", false, "Prints the help message")
 )
 
 func main() {
-	var (
-		executionConfigFilePath    = flag.String("execution-config", "config.json", "execution config file location")
-		sourceOperationFoldersPath = flag.String("operations", "operations", "source operations folder location")
-		plansOutPath               = flag.String("plans", "plans", "output plans folder location")
-		operationFilterFilePath    = flag.String("filter", "", "operation filter file location which should contain file names of operations to include")
-		timeout                    = flag.String("timeout", "30s", "timeout")
-		concurrency                = flag.Int("concurrency", 0, "how many query plan run concurrently")
-		help                       = flag.Bool("help", false, "Prints the help message")
-	)
-
 	flag.Parse()
 
 	if *help {
@@ -25,7 +25,7 @@ func main() {
 		return
 	}
 
-	err := cmd.PlanGenerator(cmd.QueryPlanConfig{
+	err := plan_generator.PlanGenerator(plan_generator.QueryPlanConfig{
 		ExecutionConfig: *executionConfigFilePath,
 		SourceDir:       *sourceOperationFoldersPath,
 		OutDir:          *plansOutPath,
